Simplified saving of PluginDialogTab configuration.
Before the configuration value for plugin dir was saved only when the value
was differet from PluginManager::pluginDirectory(). The modified state is
now semi-controlled by AbstractConfigurationDialog, which doesn't allow
saving when no edited() signal was received. So the explicit check of
plugin dir is not needed now (and with more configurable options it will not
be possible to check everything explicitly).