Thread-safe access to TileModel instance.
Instance pointer can be obtained with MainWindow::lockTileModelForRead()
(returns const pointer) or MainWindow::lockTileModelForWrite(). After end
of tile model usage the lock should be unlocked with
MainWindow::unlockTileModel().

Implementation is internally done via QReadWriteLock, which means that
all threads can access the model for reading at the same time, but write
access is exclusive. Before that it was locked with mutexes, which were
always exclusive, so it should be now faster without unnecessary lockups.