Cache configuration UI.
Original configuration split into two tabs, one for main configuration
and one for cache. Cache configuration can be slow, thus all operations
are done in separate thread and the dialog indicates that an operation
is in progress.

Only cache dir and cache plugin is stored in configuration, cache size
and block size stores the cache itself. Default values for cache size
and block size are used only when new cache is set up (the plugin name
changes or the dir changes). These values are on cache initialization
saved as default (and used in future as default for new caches).